// src/common.d.ts
export interface returnCommonPropsPromise {
  status: boolean;
  msg: string;
  tables?:any[]
}

// export interface returnGetDataSqliteT extends Partial<Pick<returnCommonPropsPromise, 'msg'>>, Omit<returnCommonPropsPromise, 'msg'>{
export interface returnGetDataSqliteT extends returnCommonPropsPromise{
  values:[] | {}[];
}
export interface returnGetDataSLStoreT extends returnCommonPropsPromise{
  data: {[key: string]: any}
}

export interface SelectWhere{
  sql?: string;
}



export interface Params {
  where?:object;
  whereKey?: {[key:string]: string[] | undefined};
  ignoreWhere?: {[key:string]: string[] | undefined};
  condition?: 'AND' | 'OR';
  stringWhere?:string | undefined;
}

export interface UpdateWhere extends Omit<Params, 'whereKey' | 'ignoreWhere'>{ }

export interface setDataSqliteOptions {
  isCreateDate?: boolean | undefined;
  rewriteTable?: boolean | undefined;
}
// ...
export interface getItemsI {
  ignoreKeys: string[]
}
// export type commonMethodSqliteT = (nameTable:string) => Promise<returnCommonPropsPromise> 
export type getItemT = (key:string) => Promise<returnGetDataSLStoreT> 
export type getItemsT = (options?:getItemsI) => Promise<returnGetDataSLStoreT> 

export type setItemT = (key:string, value: any, options?:setDataSqliteOptions) => Promise<returnCommonPropsPromise> 
// export type updateDataT = (nameTable:string, payload: object, {where, condition, stringWhere}: UpdateWhere) => Promise<returnCommonPropsPromise> 
export type remoteItemT = ({keys, ignoreKeys}:{keys?:string[], ignoreKeys?:string[]}) => Promise<returnCommonPropsPromise> 
export type dropAllData = () => Promise<returnCommonPropsPromise> 